Configured GeoWave drivers to run as plugins.
