+++ cmake/boost.cmake
@@ -13,7 +13,7 @@ if(WIN32)
     endif(MSVC)
 endif(WIN32)
 
-find_package(Boost 1.53 REQUIRED COMPONENTS program_options iostreams filesystem system)
+find_package(Boost 1.53 REQUIRED COMPONENTS program_options iostreams filesystem system thread)
 set_package_properties(Boost PROPERTIES TYPE REQUIRED
     PURPOSE "Boost provides fundamental functionality for PDAL")
 if(Boost_FOUND)
