+++ io/CMakeLists.txt
@@ -5,6 +5,5 @@ add_subdirectory(qfit)
 add_subdirectory(sbet)
 add_subdirectory(text)
 add_subdirectory(terrasolid)
-add_subdirectory(geowave)
 
 set(PDAL_TARGET_OBJECTS ${PDAL_TARGET_OBJECTS} PARENT_SCOPE)
