+++ plugins/geowave/io/GeoWaveWriter.hpp
@@ -35,6 +35,7 @@
 #pragma once
 
 #include <pdal/Writer.hpp>
+#include <pdal/StageFactory.hpp>
 
 #include <vector>
 #include <string>
@@ -45,8 +46,9 @@ namespace pdal
     class PDAL_DLL GeoWaveWriter : public Writer
     {
     public:
-        SET_STAGE_NAME("writers.geowave", "Geowave Writer")
+        SET_STAGE_NAME("writers.geowave", "GeoWave Writer")
         SET_STAGE_LINK("http://pdal.io/stages/drivers.geowave.writer.html")
+        SET_PLUGIN_VERSION("1.0.0")
 
         static Options getDefaultOptions();
